Some refactoring, removed inline, moved namespace to own file, renamed
