+++ IngredientUtils.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <string>
+
+#include "Ingredients.h"
+
+namespace IngredientUtils {
+	std::string GetIngredientName(EIngredients ingredient) {
+		return kIngredientName.at(ingredient);
+	}
+}
