+++ Ingredients.h
@@ -1,7 +1,7 @@
-#include <iostream>
+#pragma once
+
 #include <string>
 #include <map>
-#pragma once
 
 enum class EIngredients {
 	FLOUR,
@@ -10,18 +10,9 @@ enum class EIngredients {
 	EGG
 };
 
-const std::map <EIngredients, std::string> kIngredientName
-{
+const std::map<EIngredients, std::string> kIngredientName {
 	{EIngredients::FLOUR, "Faina"},
 	{EIngredients::SUGAR, "Zahar"},
 	{EIngredients::BUTTER, "Unt"},
 	{EIngredients::EGG, "Ou(a)"}
 };
-
-namespace Ingredient {
-	inline std::string GetName(EIngredients Ingredient) {
-
-		return kIngredientName.at(Ingredient);
-	}
-}
-
