+++ Recipe.cpp
@@ -1,6 +1,5 @@
 #include "Recipe.h"
-
-
+#include "IngredientUtils.h"
 
 CRecipe::CRecipe(std::string name, std::vector<SIngredientQuantity> ingredientQuantities) {
 	mName = name;
@@ -14,7 +13,7 @@ CRecipe::CRecipe(std::string name, std::vector<SIngredientQuantity> ingredientQu
 std::string CRecipe::ToString() {
 	std::string recipeText = mName + ": \n\n";
 	for (unsigned int i = 0; i < mIngredientQuantities.size(); i++) {
-		recipeText += " - " + Ingredient::GetName((mIngredientQuantities.at(i).mIngredient)) + ": "
+		recipeText += " - " + IngredientUtils::GetIngredientName((mIngredientQuantities.at(i).mIngredient)) + ": "
 			+ std::to_string(mIngredientQuantities.at(i).mQuantity) + " " + mIngredientQuantities.at(i).mUnit;
 	}
 	recipeText += "\n";
