+++ Recipe.h
@@ -1,23 +1,23 @@
+#pragma once
+
 #include <vector>
+
 #include "Ingredients.h"
-
-
 
 struct SIngredientQuantity {
 	EIngredients mIngredient;
-	/*unsigned*/double mQuantity;
+	double mQuantity;
 	std::string mUnit;
 };
 
 class CRecipe {
+public:
+	CRecipe(std::string name, std::vector<SIngredientQuantity> ingredientQuantities);
+	
+	std::string ToString();
+
 private:
 	std::string mName;
 	std::vector<SIngredientQuantity> mIngredientQuantities;
-
-public:
-	CRecipe(std::string name, std::vector<SIngredientQuantity> ingredientQuantities);
-
-	std::string ToString();
-
 
 };
